threading and sensor value output
